// frontend/generated/models/ContractStatusDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Freelance Platform API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const ContractStatusDto = {
    Unsigned: 'UNSIGNED',
    Signed: 'SIGNED',
    Terminated: 'TERMINATED'
} as const;
export type ContractStatusDto = typeof ContractStatusDto[keyof typeof ContractStatusDto];


export function instanceOfContractStatusDto(value: any): boolean {
    for (const key in ContractStatusDto) {
        if (Object.prototype.hasOwnProperty.call(ContractStatusDto, key)) {
            if (ContractStatusDto[key as keyof typeof ContractStatusDto] === value) {
                return true;
            }
        }
    }
    return false;
}

export function ContractStatusDtoFromJSON(json: any): ContractStatusDto {
    return ContractStatusDtoFromJSONTyped(json, false);
}

export function ContractStatusDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): ContractStatusDto {
    return json as ContractStatusDto;
}

export function ContractStatusDtoToJSON(value?: ContractStatusDto | null): any {
    return value as any;
}

export function ContractStatusDtoToJSONTyped(value: any, ignoreDiscriminator: boolean): ContractStatusDto {
    return value as ContractStatusDto;
}